requirements.txt
opencv-python>=4.10.0.84
mediapipe>=0.10.14
tensorflow>=2.17.0
numpy>=1.26.4
pandas>=2.0.0
scikit-learn>=1.5.1
pillow>=10.4.0
pygame>=2.5.0
